Add autorouting and bulk elevation API
This change adds autorouting to the Leaflet map so that the polylines
follow the road.

This change also adds a bulk elevation API which the front end now uses
to gather elevation data along the road.

Finally, this change refactors the front end a bit to use Leaflet's LatLng
object and its helper methods.