Subdivide routes produced by OSRM, fix routing bug
This change adds a function which subdivides the routes produced by OSRM
so that it takes full advantage of the resolution of the data on
straight roads and curved roads.

This change also fixes a bug with the auto-routing line mode where if
multiple routes were produced, multiple routes were rendered.